Fix TestNG test cases
