+++ e2e/testng-firefox.xml
@@ -6,6 +6,7 @@
         <parameter name="browser" value="firefox"/>
         <classes>
             <class name="org.yorkxin.copyasmarkdown.e2e.ContextMenuTest"/>
+            <class name="org.yorkxin.copyasmarkdown.e2e.CustomFormatPageTest"/>
             <class name="org.yorkxin.copyasmarkdown.e2e.OptionsPageTest"/>
             <class name="org.yorkxin.copyasmarkdown.e2e.PermissionsPageTest"/>
             <class name="org.yorkxin.copyasmarkdown.e2e.keyboardshortcut.OnPageContentsTest"/>
